+++ src/abi/mod.rs
@@ -11,6 +11,7 @@
 #![cfg_attr(feature = "none", doc = "`none`")]
 #![cfg_attr(feature = "dcre", doc = ", `dcre`")]
 #![cfg_attr(feature = "rstr_task", doc = ", `rstr_task`")]
+#![cfg_attr(feature = "messagebuf", doc = ", `messagebuf`")]
 //! </i>
 #![allow(non_camel_case_types)]
 #![allow(unused_imports)]