+++ src/messagebuffer.rs
@@ -1,0 +1,12 @@
+//! Message buffers (TODO)
+
+// TODO: snd_mbf
+// TODO: psnd_mbf
+// TODO: tsnd_mbf
+// TODO: rcv_mbf
+// TODO: prcv_mbf
+// TODO: trcv_mbf
+// TODO: ini_mbf
+// TODO: ref_mbf
+// TODO: acre_mbf
+// TODO: del_mbf